+++ src/modules/hooks/emoji-react/index.ts
@@ -2,7 +2,7 @@ import { boundMethod } from 'autobind-decorator';
 import { Message } from 'discord.js';
 import { parse } from 'twemoji-parser';
 
-import type { queryMessage } from '@/types.js';
+import type { queryMessage } from '@/types/types.js';
 import { Na2Client } from '@/client.js';
 
 type wordReactType = Array<[RegExp, string]>;
